 backend/requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.8.1
pinecone-client==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
EOF